requirements.txt
# Web framework for health checks
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Browser automation
playwright>=1.40.0

# Environment variables
python-dotenv>=1.0.0

# Telegram Bot
python-telegram-bot>=20.7
nest-asyncio>=1.5.8
